add aws package
