+++ aws/README.md
@@ -1,0 +1,6 @@
+# Titan AWS
+
+Amazon Web Services support for Titan. Utilized AWS services are:
+
+* Data Storage: DynamoDB
+* Message Queue: ElastiCache - Redis
