+++ aws/db_dynamodb.go
@@ -1,0 +1,1 @@
+package aws
