+++ aws/queue_elasticache.go
@@ -1,0 +1,1 @@
+package aws
